Volcano Plot

Draw a volcano plot. Currently supported data types:

glystats_ttest_res: Result from glystats::gly_ttest().

glystats_wilcox_res: Result from glystats::gly_wilcox().

glystats_limma_res: Result from glystats::gly_limma(). Only two-group comparisons are supported.

glyexp_experiment: Experiment created by glyexp::experiment(). Differential expression analysis is first performed using glystats::gly_ttest(), glystats::gly_wilcox(), or glystats::gly_limma() (controlled by the test argument), then the result is plotted.

Usage

plot_volcano(x, log2fc_cutoff = 1, p_cutoff = 0.05, p_col = "p_adj", ...)

# S3 method for class 'glyexp_experiment'
plot_volcano(
  x,
  log2fc_cutoff = 1,
  p_cutoff = 0.05,
  p_col = "p_adj",
  test = "limma",
  stats_args = list(),
  ...
)

# S3 method for class 'glystats_ttest_res'
plot_volcano(x, log2fc_cutoff = 1, p_cutoff = 0.05, p_col = "p_adj", ...)

# S3 method for class 'glystats_wilcox_res'
plot_volcano(x, log2fc_cutoff = 1, p_cutoff = 0.05, p_col = "p_adj", ...)

# S3 method for class 'glystats_limma_res'
plot_volcano(
  x,
  log2fc_cutoff = 1,
  p_cutoff = 0.05,
  p_col = "p_adj",
  contrast = NULL,
  ...
)

Arguments

x

An object to be plotted.

log2fc_cutoff

The log2 fold change cutoff. Defaults to 1.

p_cutoff

The p-value cutoff. Defaults to 0.05.

p_col

The column name for p-value. Defaults to "p_adj". Can also be "p_val" (raw p-values without multiple testing correction).

...

Other arguments passed to EnhancedVolcano::EnhancedVolcano().

test

"ttest", "wilcox", or "limma". Default is "limma".

stats_args

A list of keyword arguments to pass to glystats::gly_ttest(), glystats::gly_wilcox(), or glystats::gly_limma().

contrast

A character string specifying the contrast to plot, in the format of "group1_vs_group2". Must be one of the contrasts in the result. When there is only one contrast (two-group comparison), it can be NULL (default).

Value

A ggplot object.